refactor: new test structure robot model
