+++ source/robot_model/test/tests/fixtures/generateRobotModelTestConfigurations.m
